requirements.txt
openai>=1.10.0
python-dotenv>=1.0.0
mcp>=0.2.0